Create Python_automation.py
This is the Latest python script that I have worked on and decided to push my code to github. 
I have written a python automation to monitor oracle databases for various transactions on daily basis. it was a manual process previously like logging into database, run set of queries against different databases and manually review the results for any violations and suspicious activities etc. it used to take lot of time when we run the queries and getting the results manually so decided to automate this manual process. My script will login to several databases and generate CSV logs of raw output as well as perform basic analysis